requirements.txt
requests
PyGithub
python-dotenv
openai
PyYAML
Flask
